fastapi-service/pyproject.toml
[project]
name = "ollama-discord-fastapi"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "aioboto3>=15.5.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "strands-agents>=1.18.0",
    "strands-agents-tools>=0.2.17",
    "ollama>=0.4.4",
    "openai>=1.0.0",
    "tiktoken>=0.5.1",
    "python-dotenv>=1.0.0",
    "websockets>=12.0",
    "aiohttp>=3.9.0",
    "ddgs>=6.3.5",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "langchain>=0.1.0",
    "langchain-text-splitters>=0.0.1",
    "pypdf>=5.1.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
    "moto[dynamodb]>=4.2.0",
]
